requirements.txt
absl-py==0.10.0
alabaster==0.7.12
astor==0.8.1
astroid==2.4.2
attrs==20.2.0
autopep8==1.5.4
avro==1.10.0
Babel==2.8.0
backcall==0.2.0
blis==0.4.1
boto3==1.16.1
botocore==1.19.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
conllu==4.2.1
contextlib2==0.6.0.post1
coverage==5.3
cymem==2.0.3
decorator==4.4.2
docopt==0.6.2
docutils==0.16
editdistance==0.5.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
es-core-news-sm @ file:///home/lzk/baseline_for_naacl/Span_OIE/es_core_news_sm-2.3.1.tar.gz
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Cors==3.0.9
ftfy==5.8
future==0.18.2
gast==0.4.0
gensim==3.8.3
gevent==20.9.0
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.32.0
h5py==2.10.0
hyperopt==0.1.2
idna==2.10
imagesize==1.2.0
importlib-metadata==2.0.0
iniconfig==1.1.1
ipdb==0.13.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.4
itsdangerous==1.1.0
jedi==0.17.2
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json-tricks==3.15.3
jsonnet==0.16.0
jsonpickle==1.4.1
Keras==2.2.4
Keras-Applications==1.0.8
keras-contrib @ git+https://www.github.com/keras-team/keras-contrib.git@3fc5ef709e061416f4bc8a92ca3750c824b5d2b0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3
MarkupSafe==1.1.1
mccabe==0.6.1
murmurhash==1.0.2
netifaces==0.10.9
networkx==2.5
nltk==3.5
nni==1.9
numpy==1.19.2
numpydoc==1.1.0
overrides==3.1.0
packaging==20.4
pandas==1.1.3
parsimonious==0.8.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.6.1
pkuseg==0.0.22
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.3
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.7.1
pylint==2.6.0
pyltp==0.4.0
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
PythonWebHDFS==0.2.3
pytorch-crf==0.7.2
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.9.27
requests==2.24.0
responses==0.12.0
rope==0.17.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sacremoses==0.0.43
schema==0.7.3
scikit-learn==0.23.2
scipy==1.5.2
sentencepiece==0.1.91
seqeval==1.2.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smart-open==3.0.0
snowballstemmer==2.0.0
spacy==2.3.2
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
srsly==1.0.2
tensorboard==1.14.0
tensorboardX==2.1
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
thinc==7.4.1
threadpoolctl==2.1.0
tokenizers==0.9.2
toml==0.10.1
torch==1.3.0
tqdm==4.50.2
traitlets==5.0.5
transformers==3.4.0
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.10
wasabi==0.8.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wget==3.2
word2number==1.1
wrapt==1.12.1
zenodo-get==1.3.2
zipp==3.2.0
zope.event==4.5.0
zope.interface==5.1.2
